refactor: split contextual schedule planner
